Change "VERSION" created by Windows build to "SILO_VERSION" to remove collision with "VERSION" defined in hdf5. Remove define of _HDF5USEDLL_ by the build, as it is predefined by hdf5. Add define of LITE_EXPORTS to silo project so pdb_lite symbols are properly exported. Make each project output a build log with its name (eg browser_BuildLog instead of just BuildLog) Have silex copy QT dependencies into output dir so it will run on Windows. Add inlined argument definitions to filter functions. Added pragma to disable certain warnings in fpzip code. Removed a few unreferenced variables. Fixed a few signed/unsigned comparison warnings by casting to int.
git-svn-id: file:///usr/gapps/silo/svn/trunk@310 92378438-2262-0410-827d-edb41c1c42c8

+++ src/silo/silo.c
@@ -9403,7 +9403,7 @@ _DBstrprint(FILE *fp, char *strs[], int nstrs, int order, int left_margin,
     maxwidth = strlen(sorted_strs[0]);
 
     for (i = 1; i < nstrs; i++) {
-        maxwidth = MAX(maxwidth, strlen(sorted_strs[i]));
+        maxwidth = MAX(maxwidth, (int)strlen(sorted_strs[i]));
     }
 
      /*----------------------------------------
